move vars to component_readiness
