+++ sippy-ng/src/App.js
@@ -1,5 +1,5 @@
 import './App.css'
-import { CompReadyVarsProvider } from './CompReadyVars'
+import { CompReadyVarsProvider } from './component_readiness/CompReadyVars'
 import { createTheme, makeStyles, useTheme } from '@material-ui/core/styles'
 import { CssBaseline, Grid, MuiThemeProvider } from '@material-ui/core'
 import { getReportStartDate, relativeTime } from './helpers'
